Fix argument parsing (#8)
Finally found a way to reliably trick OptionParser into behaving.

It turns out it doesn't just throw away non option arguments on error,
it throws away everything on error. So save it, remove the offending
option, and restore.

This reliably keeps exactly the right options in all of my testing.