Fix enterprise github
In order to fix option parsing in #22 we moved the creation of the
`SugarJar::Commands` object to before we finished parsing options.

While that's necessary, we need to re-create it *after* option parsing
for options to have an effect (obviously).

This fixes that. It also fixes the default GH user setting and adds
a bit of debugging.

Signed-off-by: Phil Dibowitz <phil@ipom.com>