Always use SSH for forked remote; handle various repo paths better
A user who clones-and-forks an https repo they don't have commit
access to, still wants their forked repo to be using SSH so they
can push to it and make PRs.

In addition, we were misparsing repo paths in a variety of different
ways, so this centralizes that parsing into two methods, uses them
consistently, and then unittests them in a variety of cases.

This closes #56

Signed-off-by: Phil Dibowitz <phil@ipom.com>