Add smartpr; add dirty checks in {s,f}push
Add a new `smartpullrequest` command (aliases: `spr` and `smartpr`).
It's a shortcut for `hub pull-request`, but checks for a dirty repo
first.

That closes #51

Since we have a dirty check now, this adds it to `smartpush` and
`forcepush` as well as refactors those to DRY them up a bit.

Finally, this adds `--ignore-dirty` and `--ignore-prerun-failure` to
allow bypassing botht he new dirty check and the existing prerun checks.

Signed-off-by: Phil Dibowitz <phil@ipom.com>