Support 'main' as a primary branch name (#93)
Since a lot of what Sugar Jar does depends on finding the primary
branch, 'master' was hard-coded a lot. This undoes that hard-coding and
should allow it to seemlessly work regardless of if the primary branch
is `master` or `main`. To be primarily future-compatible as well as to
honor the spirit of the move away from `master` terminology, it tries to
find `main` and uses it if it exists, and then falls back to `master`.

Signed-off-by: Phil Dibowitz <phil@ipom.com>