Improve parsing for subcommands
In #89, @infcted exposed a bug in our parsing. We check to see if any
naked arguments in the command-line are valid SJ subcommands, and one
could have that, but it be an arg to an option, and then when we figure
out our real sub-command, it turns out it's not valid.

This changes that to actually validate our subcommand is valid once we
have it so that this doesn't happen.

Closes #89

Signed-off-by: Phil Dibowitz <phil@ipom.com>