Fix rebase issues
Rebase failures often ended up in confusing messages that don't tell
users how to get out of the mess. This catches those errors and gives
users pointers.

If it fails in `sj up`, we leave the repo as-is and give the user both
the revert option or the fix option and all the output:
```
$ sj up
FATAL: cookbook_for_mac: Failed to rebase on upstream/master. Leaving the repo as-is. You can get out of this with a `git rebase --abort`. Output from failed rebase is:
STDOUT:
	First, rewinding head to replay your work on top of it...
	Applying: Mucking about
	Using index info to reconstruct a base tree...
	M	cookbooks/<redacted>/recipes/macos.rb
	Falling back to patching base and 3-way merge...
	Auto-merging cookbooks/<redacted>/recipes/macos.rb
	CONFLICT (content): Merge conflict in cookbooks/<redacted>/recipes/macos.rb
	Patch failed at 0001 Mucking about
	Use 'git am --show-current-patch' to see the failed patch

	Resolve all conflicts manually, mark them as resolved with
	"git add/rm <conflicted_files>", then run "git rebase --continue".
	You can instead skip this commit: run "git rebase --skip".
	To abort and get back to the state before "git rebase", run "git rebase --abort".

STDERR:
	error: Failed to merge in the changes.

```

In up-all, we do as before, but let the user know to try `sj up`:
```
$ sj upall
ERROR: cookbook_for_mac failed rebase. Reverting attempt and moving to next branch. Try `sj up` manually on that branch.
firewall rebased on upstream/master
```

Closes #88

Signed-off-by: Phil Dibowitz <phil@ipom.com>