Implement check lister-commands
It is not uncommon to want to discover the linters or unit tests to be
run dynamically. A common example is deciding the linters based on the
files that were modified.

This PR implements two new configs in the Repository Configuration:
`lint_list_cmd` and `unit_list_cmd` which are a command to run which
will return the linters/unit tests to run.

This also moves our linters to `lint_list_cmd`.

Closes #80.

Signed-off-by: Phil Dibowitz <phil@ipom.com>