Lots of breanch-cleanup work
The cleanup logic has gotten messed up at some point, so bclean{all},
always checked what the _current_ branch was tracking instead of what
the branch-to-be-cleaned was tracking. Fortunately, this almost
certainly lead to lost of false-negatives instead of false-positives,
but it's still a bummer. This fixes tha.

In addition this adds some new commands for cleaning up remote branches
in addition to local branches and renamed some commands for clarity:

* `localbranchclean` / `lbclean` - local branch clean. aliased as
  `bclean` for back-comat
* `localbranchcleanall` / `lbcleanall` - local all branch clean aliased
  as `bcleanall` for back-compat
* `remotebranchclean` / `rbclean` - remote branch clean
* `remotebranchcleanall` / `rbcleanall` - remote all branch clean
* `globalbranchclean` / `gbclean` - local+remote branch clean
* `globalbranchcleanall` / `gbcleanall` - local+remote all branch clean

Closes #199

Signed-off-by: Phil Dibowitz <phil@ipom.com>